# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/index": {
            "get": {
                "tags": [
                    "BrandSetting"
                ],
                "summary": "Get list of brandSettings",
                "description": "Returns list of brandSettings",
                "operationId": "getBrandSettingList",
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/brandSave/{id}": {
            "post": {
                "tags": [
                    "BrandSetting"
                ],
                "summary": "Update existing BrandSetting",
                "description": "Returns updated BrandSetting data",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "BrandSettings id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "pdf_logo": {
                                        "type": "file",
                                        "format": "file"
                                    },
                                    "app_logo": {
                                        "type": "file",
                                        "format": "file"
                                    },
                                    "favicon": {
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBrandSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/url-save/{id}": {
            "put": {
                "tags": [
                    "BrandSettingsUrl"
                ],
                "summary": "Update BrandSettings url in DB",
                "description": "Update BrandSettings url in DB",
                "operationId": "urlstore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of BrandSettings",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "privacy_policy_url",
                                    "terms_and_conditions_url",
                                    "legal_disclaimer_url",
                                    "knowledge_base_url",
                                    "faq_url",
                                    "service_docs_url",
                                    "service_tutorials_url",
                                    "product_api_url",
                                    "api_code_github_repo_url",
                                    "ask_a_question_url",
                                    "ssh_key_generation_tutorial_url",
                                    "coupon_code_terms_and_conditions_url",
                                    "referral_terms_and_conditions_url",
                                    "facebook_url",
                                    "twitter_url",
                                    "object_storage_s3cmd_doc_url"
                                ],
                                "properties": {
                                    "privacy_policy_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "terms_and_conditions_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "legal_disclaimer_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "knowledge_base_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "faq_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "service_docs_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "service_tutorials_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "product_api_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "api_code_github_repo_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "ask_a_question_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "ssh_key_generation_tutorial_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "coupon_code_terms_and_conditions_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "referral_terms_and_conditions_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "facebook_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "twitter_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    },
                                    "object_storage_s3cmd_doc_url": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "http://test/phpmyadmin/"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateBrandSettingRequest": {
                "title": "Profile Update Request",
                "description": "Update User Profile request body",
                "required": [
                    "app_name",
                    "theme_color",
                    "secondary_color"
                ],
                "properties": {
                    "app_name": {
                        "title": "app_name",
                        "description": "App name",
                        "type": "string",
                        "example": "StackConsole2.0-Laravel"
                    },
                    "theme_color": {
                        "title": "theme_color",
                        "description": "theme_color",
                        "type": "string",
                        "example": "Red"
                    },
                    "secondary_color": {
                        "title": "secondary_color",
                        "description": "secondary_color",
                        "type": "string",
                        "example": "Red"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "BrandSetting",
            "description": "API Endpoints of BrandSetting"
        }
    ]
}